Launch report from completion page and single page activity, add showReport util function
