+++ src/components/single-page/submit-button.tsx
@@ -1,16 +1,13 @@
 import React from "react";
+import { showReport } from "../../utilities/report-utils";
 
 import "../activity-page/bottom-buttons.scss";
 
-interface SubmitButtonProps {
-  onSubmit?: () => void;
-}
-
-export const SubmitButton: React.FC<SubmitButtonProps> = (props) => {
+export const SubmitButton: React.FC = () => {
   return (
     <div className="bottom-buttons center">
       <div>
-        <button className="button" onClick={props.onSubmit} data-cy="submit-button">Submit</button>
+        <button className="button" onClick={showReport} data-cy="submit-button">Submit</button>
       </div>
     </div>
   );
